WP/I18n: fix `SuperfluousDefaultTextDomain` auto-fix code to allow for named params
The auto-fixer for the `SuperfluousDefaultTextDomain` warning attempts to remove a complete parameter, though it will bow out from auto-fixing when there is a comment in the tokens which would be removed.

This auto-fixer code and the code to determine whether an auto-fix is possible, now needs adjusting to allow for function calls with named parameters as otherwise the fixer could cause parse errors.
* The parameter label and the `:` also needs to be removed.
* If the named parameter is the first parameter, there will not be a preceding comma, but we will need to remove the comma behind it.
* The range of tokens which needs to be checked for comments to determine whether the error is auto-fixable needs to be expanded.

This commit makes those changes.

Includes additional unit tests to safeguard the fixes.